Compute sri hashes and compare with downloaded ones.
The sri hashes of the current worker files are stored in a
file "srt.txt". This file is updated each time the worker
runs, and it can also be updated manually with

python worker.py --only_config

The worker will also download the file (currently not existing)

https://raw.githubusercontent.com/glinscott/fishtest/master/worker/sri.txt

and check if the hashes in that file coincide with the sri hashes
of the current worker.

If not then a message is printed on the terminal.
The worker is not stopped but the flag worker_info['modified'] is
set to True. In a subsequent PR the server will use this information
in the event log messages (for debugging purposes).

GitHub actions are used to check if the file srt.txt matches
the worker files during pushes or pull requests.