Fixup commit
This is a PR on top of #2380.

We offer the user the option to make an empty commit to their
dev branch, with message equal to the pr message. When
rebasing/squashing this commit can be used with "f -C".

Also: use fetch with cache: "no-store". This resolves a lot
of confusing behaviour and we are caching things internally
anyway.

Future developments:

Rebase and squash the user's repo ourselves - in the absence
of conflicts. This is not particularly difficult but
sufficient safeguards need to be in place so that we
do not destroy the user's work if something goes wrong,